GUI update
1. In the payment panel, when selecting a payment type from the drop
down, an amount label or credit card label will be displayed depending
on what is chosen.

2. The Payment text field is disabled by default. It is only enabled
after the "Enter" name button is pressed.

3. In the payment panel, the "Submit Order" button is only enabled if
the Payment text field is not empty. Also, "Submit Order" button now
updates the time displayed.

4. After inputting a name in the customer name text field and pressing
"Enter", the name will be displayed in the Invoice label. For example if
Ray Wong was entered, the Invoice label will be updated to display
"Invoice   for   Ray Wong". This is to show the user the button did
something.